Created a separate `utils` directory for code that might be common to more than one program.
